Fix PageStore::root
Since PageStore::root is not only used to get the root of the current page, but also of “future” page with a changed slug, the cached root can only be used if no `$page` parameter is passed.

In other cases, `$page->dirname()` should be used to get correct results. Otherwise the `$page` parameter itself doesn’t make sense at all.