Data: Fix handling of corrupted data
If a file can’t be read as the contents are corrupted, we can’t return an array. I have opted to return null in this case.

There is one issue: Nullable return types (the question mark in front of the return type) were introduced in PHP 7.1. If we need support for PHP 7.0, we need to remove that return type declaration entirely.