Fix session filesystem error on Windows
Apparently Windows doesn’t allow the deletion of files that are still open. We needed to keep the file open until the file was deleted to ensure thread-safeness (otherwise another process could have written to the file right after we deleted it).

Thread-safeness is now ensured via a different method that will be compatible with Windows servers.