Include all PHP files in CS fixer
I think it makes sense to also ensure a consistent coding style in the config files as well as the tests.

I’ve checked all changed files. In some places the fixer got rid of manual indenting in tests, but I think that’s fine. I’m sending this as a PR anyway so that you can decide.