API: Respect configured locale on single-lang site
The `$kirby->setCurrentLanguage()` method not only sets the active language, but also the active PHP locale. It should be called for every API request, no matter if a language is specified or not.

The method will automatically fall back to the default language and its locale – or on single-lang sites to the configured `locale` option.