App::$nestIgnoreOptions; fix options regression
This fixes a regression introduced by #2572:

Because all options are now nested as much as possible, hooks registered with the `hooks` option would also get nested because the event names and therefore the array keys inside the `hooks` array contain dots.

The `hooks` option is now blacklisted from nesting. If we find out that more options are affected, we can add those to `AppPlugin::$nestIgnoreOptions` as well.