Support for old plugin option syntax #2697
As the plugin option defaults are now registered as one array for each plugin and not as one array key per option, options set using the old `vendor.plugin.option` syntax in the site config wouldn't have an effect.

These options are now automatically converted to the new structure once the App instance is fully loaded.