Open dropdown when clicking on empty files field
Just had a client have trouble uploading files from the files field because clicking on the empty field opened the file picker and there was no image there. He didn't notice the **Add** button on the right and it took some back and forth to help him with that.

I noticed the `Files` component had an unused `prompt` method that opens the dropdown when upload is enabled. This PR puts that method to good use to improve the UX on this component.