Dom: Always collect all elements from iterators
By preserving the keys (default behavior of `iterator_to_array()`), elements with the same name would overwrite each other in the generated array. This causes issues for example with attributes/elements with and without namespace but with the same local name. As we don't use the keys, we can ignore them.